refactor: extract PDF conversion tests
